Fixed issue on Linux where prepareStorage() hangs
On some linux distros, mockgoose would download a wrong version of
mongodb, which dont print the expected line on startup (these are
development versions of mongodb and shouldnt be used at all). This led
prepareStorage() to hang indefinitely.

The fix is setting the exact version we want to download to 3.2.1... at
least for now it works.

More details on this issue: https://github.com/Mockgoose/Mockgoose/issues/36#issuecomment-309678249